requirements/dev.txt
-r base.txt
django-autofixture==0.12.1
django-debug-toolbar==1.9.1
factory-boy==2.10.0
Faker==0.8.11
flake8-docstrings==1.3.0
flake8==3.5.0
freezegun==0.3.10
isort==4.3.4
pytest-cov==2.5.1
pytest-django==3.1.2
pytest-factoryboy==2.0.1
pytest==3.4.2
